refactor(interpreter): break down interpreter (#110)
* refactor(interpreter): break down interpreter

* refactor(interpreter): refactor function names
