+++ src/CMakeLists.txt
@@ -6,7 +6,6 @@ set(HEADERS
     snuk_string.h
     lexer.h
     darray.h
-    interpreter.h
     runtime.h
 )
 
@@ -17,7 +16,6 @@ set(SRCS
     io.c
     lexer.c
     darray.c
-    interpreter.c
     runtime.c
 )
 
@@ -26,3 +24,4 @@ target_include_directories(snuk PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
 target_sources(snuk PRIVATE ${HEADERS} ${SRCS})
 
 add_subdirectory(parser)
+add_subdirectory(interpreter)
