+++ src/runtime.c
@@ -14,9 +14,9 @@ void snuk_runtime_execute(Runtime *rt, const char *src) {
         // snuk_parser_log_item(item);
         // log_trace("", NULL);
         SnukValue value = snuk_interpreter_exec_item(&rt->interpreter, item);
-        snuk_interpreter_log_value(value);
+        snuk_value_log(value);
         log_trace("", NULL);
-        snuk_interpreter_free_value(value);
+        snuk_value_free(value);
     }
 
     snuk_parser_deinit(&parser);
